backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai>=1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pillow==10.1.0
pytesseract==0.3.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
httpx==0.25.2
aiofiles==23.2.1
email-validator==2.1.0
stripe==7.5.0
plaid-python==9.8.0

